+++ tests/test_summary.py
@@ -588,7 +588,7 @@ class SummaryTest2(CoverageTest):
 
         report = repout.getvalue().replace('\\', '/')
         report = re.sub(r"\s+", " ", report)
-        self.assertIn("tests/modules/pkg1/__init__.py 1 0 100%", report)
+        self.assertIn("tests/modules/pkg1/__init__.py 2 0 100%", report)
         self.assertIn("tests/modules/pkg2/__init__.py 0 0 100%", report)