+++ CHANGES.rst
@@ -15,9 +15,13 @@ Version 4.1
   are seeing corrupt data files, but this lets them continue combining anyway.
   Prompted by `issue 418`_.
 
+- Code that uses ``sys.settrace(sys.gettrace())`` used to incur more than a 2x
+  speed penalty.  Now there's no penalty at all. Fixes `issue 397`_.
+
 - Pyexpat C code will no longer be recorded as a source file, fixing
   `issue 419`_.
 
+.. _issue 397: https://bitbucket.org/ned/coveragepy/issues/397/stopping-and-resuming-coverage-with
 .. _issue 418: https://bitbucket.org/ned/coveragepy/issues/418/json-parse-error
 .. _issue 419: https://bitbucket.org/ned/coveragepy/issues/419/nosource-no-source-for-code-path-to-c